Use Int32 instead of UInt8 for stillFrameCount
To ensure overflows won't happen, even if they hardly cause any damage
  in this case... unless Udon has it's own story to tell about overflows
  I mean how knows

+++ Assets/Scripts/JanItemSync.cs
@@ -104,11 +104,11 @@ public class JanItemSync : UdonSharpBehaviour
     private const float SmallMagnitudeDiff = 0.02f;
     private const float SmallAngleDiff = 10f;
     private const float ConsistentOffsetDuration = 0.1f;
-    private const byte ConsistentOffsetFrameCount = 3;
+    private const int ConsistentOffsetFrameCount = 3;
     private float prevPositionOffsetMagnitude;
     private Quaternion prevRotationOffset;
     private float consistentOffsetStopTime;
-    private byte stillFrameCount; // to prevent super low framerate from causing false positives
+    private int stillFrameCount; // to prevent super low framerate from causing false positives
 
     // DesktopWaitingForHandToMoveState
     private const float HandMovementAngleDiff = 20f;
